Moved context namespace
