--- src/Hypzeh.Web/ClientApp/components/App/components/context/view/Context.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import { initialState } from './reducer';
-
-const Context = React.createContext(initialState);
-
-export default Context;
